requirements.txt
# requirements.txt (Updated for Web UI)

# --- Core Pipeline Dependencies ---
# Pin PyTorch/Torchaudio/NumPy
torch==2.3.1
torchaudio==2.3.1
numpy<2.0

# Core ASR and Diarization - Install from PyPI
faster-whisper>=1.0.0 # Let pip handle dependencies including ctranslate2
pyannote.audio>=3.1.1,<3.4 # Keep pinned pyannote

# Other dependencies
speechbrain>=1.0.0
faiss-cpu>=1.7.4 # Use faiss-gpu if your base image/hardware supports it and you install accordingly
transformers>=4.48.0
optree>=0.13.0

# Utilities
soundfile>=0.12.1
librosa>=0.10.0
pandas>=1.5.0
tqdm>=4.62.3
requests>=2.28.0
huggingface-hub>=0.17.0
ffmpeg-python>=0.2.0
av>=10.0.0 # Still needed by faster-whisper

cucco>=2.0.0
inflect>=7.0.0

# --- Web UI & Task Queue Dependencies ---
Flask>=2.0.0
Celery>=5.0.0
redis>=4.0.0 # For Celery broker/backend
Werkzeug>=2.0.0 # Often needed with Flask for file handling

huggingface_hub[hf_xet]

soundfile
